backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
httpx==0.27.0
pydantic-settings==2.3.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.31
celery==5.4.0
redis==5.0.6
celery==5.4.0
pyjwt==2.8.0
passlib==1.7.4
